Refactor reports controllers to separate them out.
Add new mocks helpers for jasmine tests
Update eslint to allow latest ES syntax
